@@ -28,23 +28,28 @@ dependencies:
   http:
   path:
   shared_preferences: ^0.5.8
-  location: ^3.0.2
-  provider: ^4.3.1
-  geocoder: ^0.2.1
-  google_maps_flutter: ^0.5.28+1
   intl: ^0.16.1
   file_picker: ^1.13.0
   dio: ^3.0.9
   image_picker: ^0.6.7+4
-  flutter_google_places: ^0.2.5
+
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
   cupertino_icons: ^0.1.3
-  google_maps_webservice: ^0.0.17
   dropdown_formfield: ^0.1.2
+  rflutter_alert: ^1.0.3
+  flutter_otp: ^0.3.2
+  otp: 2.2.1
+  
+  location: ^3.0.2
+  provider: ^4.3.1
+  geocoder: ^0.2.1
+  google_maps_flutter: ^0.5.28+1
+  google_maps_webservice: ^0.0.17
+  flutter_google_places: ^0.2.5
   search_map_place: ^0.3.0
   geolocator: ^5.3.2+2
-  rflutter_alert: ^1.0.3
+  razorpay_flutter: 1.1.0
 dev_dependencies:
   flutter_test:
     sdk: flutter
