@@ -1,5 +1,7 @@
 class Utils {
   static final String URL = "http://192.168.43.141//expp/";
-  static const String API_KEY = "AIzaSyBuZTVFf0pDt_MgQedl5aNsOxu286k7Wmw";
+  static const String API_KEY = "AIzaSyDjXkJePIhhPXzWuCmri3AQFEvd0jDsv7E";
+  static const String PAYMENT_KEY = "rzp_test_Wz5PhgaVMTAnAu";
+
   //static final String URL = "https://bazaar1709.000webhostapp.com/";
 }
